LayoutTests/inspector/debugger/break-on-uncaught-exception-catch-expected.txt

CONSOLE MESSAGE: line 12: testCatch
CONSOLE MESSAGE: line 7: throwing TestError
CONSOLE MESSAGE: line 16: catch TestError
CONSOLE MESSAGE: line 18: DONE
Checking pause locations when pausing on uncaught exceptions.
